code refactor #5
