+++ src/app/form-control/form-control-base.spec.ts
@@ -1,0 +1,7 @@
+import { FormControlBase } from './form-control-base';
+
+describe('FormControlBase', () => {
+  it('should create an instance', () => {
+    expect(new FormControlBase()).toBeTruthy();
+  });
+});
