+++ src/app/form-control/form-control-base.ts
@@ -1,4 +1,4 @@
-export class QuestionBase<T> {
+export class FormControlBase<T> {
     value: T;
     key: string;
     label: string;
