+++ src/app/form-control/question.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { FormControlService } from './form-control.service';
+
+describe('FormControlService', () => {
+  let service: FormControlService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(FormControlService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
