+++ src/app/form/client-view/client-view.component.ts
@@ -1,21 +1,21 @@
 import { Component, Input, OnInit }  from '@angular/core';
 import { FormGroup }                 from '@angular/forms';
 
-import { QuestionBase }              from '../question-base';
-import { QuestionControlService }    from '../question-control.service';
+import { FormControlBase }              from '../../form-control/form-control-base';
+import { FormControlService }    from '../../form-control/form-control.service';
 
 @Component({
   selector: 'app-formulario-dinamico-consumidor',
-  templateUrl: './formulario-dinamico-consumidor.component.html',
-  providers: [ QuestionControlService ]
+  templateUrl: './client-view.component.html',
+  providers: [ FormControlService ]
 })
 export class FormularioDinamicoConsumidorComponent implements OnInit {
 
-  @Input() questions: QuestionBase<string>[] = [];
+  @Input() questions: FormControlBase<string>[] = [];
   form: FormGroup;
   payLoad = '';
 
-  constructor(private qcs: QuestionControlService) {  }
+  constructor(private qcs: FormControlService) {  }
 
   ngOnInit() {
     this.form = this.qcs.toFormGroup(this.questions);
