--- src/app/pedido.spec.ts
@@ -1,7 +1,0 @@
-import { Pedido } from './pedido';
-
-describe('Pedido', () => {
-  it('should create an instance', () => {
-    expect(new Pedido()).toBeTruthy();
-  });
-});
