--- src/app/question-base.spec.ts
@@ -1,7 +1,0 @@
-import { QuestionBase } from './question-base';
-
-describe('QuestionBase', () => {
-  it('should create an instance', () => {
-    expect(new QuestionBase()).toBeTruthy();
-  });
-});
