--- src/app/question-control.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { QuestionControlService } from './question-control.service';
-
-describe('QuestionControlService', () => {
-  let service: QuestionControlService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(QuestionControlService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
